生成资源文件测试通过

+++ src/main/java/com/xiafei/tools/utils/FileUtils.java
@@ -5,6 +5,7 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -23,7 +24,14 @@ public final class FileUtils {
 
     }
 
-    public static void outPutToFileByLine(String filePath, List<String> content) {
+    /**
+     * 按行输出内容到文件
+     *
+     * @param filePath  文件完整路径（包含文件名及后缀）
+     * @param content   要输出的内容列表，列表每一个字符串代表文件一行
+     * @param isNewFile 是否替换原文件，若为否，则在源文件上追加（若存在）
+     */
+    public static void outPutToFileByLine(final String filePath, final List<String> content, final boolean isNewFile) {
         if (filePath == null || content == null || content.isEmpty()) {
             return;
         }
@@ -31,7 +39,7 @@ public final class FileUtils {
         final String regex = "^.*\\.[a-z]+$";
         if (filePath.matches(regex)) {
             final File file = new File(filePath);
-            if (file.exists()) {
+            if (isNewFile && file.exists()) {
                 file.delete();
             }
             try (FileOutputStream outputStream = new FileOutputStream(file, true);
@@ -39,19 +47,26 @@ public final class FileUtils {
                  BufferedWriter bufferedWriter = new BufferedWriter(outputStreamWriter)
             ) {
 
-                file.createNewFile();
+                if (!file.createNewFile() && !isNewFile) {
+                    bufferedWriter.write("#########################################################");
+                    bufferedWriter.newLine();
+                    bufferedWriter.write("################" + DateUtils.toString(new Date()) + "追加#################");
+                    bufferedWriter.newLine();
+                    bufferedWriter.write("#########################################################");
+                    bufferedWriter.newLine();
+                }
                 for (String line : content) {
                     bufferedWriter.write(line);
                     bufferedWriter.newLine();
                 }
                 bufferedWriter.flush();
             } catch (IOException e) {
-                throw new RuntimeException("创建文件失败，请检查文件夹是否已经建好");
+                throw new RuntimeException("创建文件失败，请检查文件夹是否已经建好" + filePath);
 
             }
 
         } else {
-            throw new RuntimeException("文件路径必须包含文件名和文件类型");
+            throw new RuntimeException("文件路径必须包含文件名和文件类型" + filePath);
         }
     }
 }
